Remove dead code
This mostly removes code left around after
21051df18e4bfa17e4689f2b6b4df6b03c0202e0 / GH PR#503, but also some
older code that is not used any longer.

`wait_until_state` function is moved to camayoc/tests/qpc/utils.py.
